back/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
flask-sqlalchemy==3.1.1
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
ollama==0.1.7
openpyxl==3.1.2
numpy==1.24.3
jieba==0.42.1

